+++ crates/extrinsics/src/instantiate.rs
@@ -31,7 +31,10 @@ use super::{
     StorageDeposit,
     TokenMetadata,
 };
-use crate::extrinsic_opts::ExtrinsicOpts;
+use crate::{
+    check_env_types,
+    extrinsic_opts::ExtrinsicOpts,
+};
 use anyhow::{
     anyhow,
     Context,
@@ -180,6 +183,7 @@ impl InstantiateCommandBuilder<state::ExtrinsicOptions> {
 
         let rpc_cli = RpcClient::from_url(&url).await?;
         let client = OnlineClient::from_rpc_client(rpc_cli.clone()).await?;
+        check_env_types(&client, &transcoder)?;
         let rpc = LegacyRpcMethods::new(rpc_cli);
 
         let token_metadata = TokenMetadata::query(&rpc).await?;