+++ crates/extrinsics/src/lib.rs
@@ -16,6 +16,7 @@
 
 mod balance;
 mod call;
+mod env_check;
 mod error;
 mod events;
 mod extrinsic_opts;
@@ -29,6 +30,7 @@ mod upload;
 mod integration_tests;
 
 use colored::Colorize;
+use env_check::compare_node_env_with_contract;
 use subxt::utils::AccountId32;
 
 use anyhow::{
@@ -368,6 +370,16 @@ pub async fn fetch_contract_info(
         }
         None => Ok(None),
     }
+}
+
+fn check_env_types<T>(
+    client: &OnlineClient<T>,
+    transcoder: &ContractMessageTranscoder,
+) -> Result<()>
+where
+    T: Config,
+{
+    compare_node_env_with_contract(client.metadata().types(), transcoder.metadata())
 }
 
 #[derive(serde::Serialize)]